allow select without table and make select more flexible

+++ src/actions.coffee
@@ -8,11 +8,17 @@ asRaw = (x) ->
     unless 'string' is typeof x
         throw new Exception 'raw or string expected'
 
-    {sql: x, params: []}
+    {
+        sql: -> x
+        params: -> []
+    }
 
 insert =
     sql: (mohair) ->
         that = this
+
+        unless mohair._table?
+            throw new Error 'sql of insert requires call to table before it'
 
         table = mohair._escape mohair._table
         keys = Object.keys(that._data)
@@ -41,6 +47,10 @@ module.exports.insert = (data) ->
 insertMany =
     sql: (mohair) ->
         that = this
+
+        unless mohair._table?
+            throw new Error 'sql of insertMany requires call to table before it'
+
         table = mohair._escape mohair._table
         first = that._array[0]
         keys = Object.keys(first)
@@ -84,7 +94,26 @@ select =
             sql += parts.join(', ')
             sql += ' '
 
-        sql += "SELECT #{that._sql} FROM #{table}"
+        sql += "SELECT "
+        parts = []
+        that._selects.forEach (s) ->
+            if isRaw s
+                parts.push '(' + s.sql() + ')'
+            else if 'object' is typeof s
+                keys = Object.keys s
+                if keys.length is 0
+                    throw new Error 'select object must have at least one property'
+                keys.forEach (key) ->
+                    value = s[key]
+                    if isRaw value
+                        parts.push '(' + value.sql() + ') AS ' + key
+                    else
+                        parts.push value + ' AS ' + key
+            else
+                parts.push s
+        sql += parts.join ', '
+        if mohair._table?
+            sql += " FROM #{table}"
         mohair._joins.forEach (join) ->
             sql += " #{join.sql}"
             sql += " AND (#{join.criterion.sql()})" if join.criterion?
@@ -102,7 +131,15 @@ select =
             Object.keys(mohair._with).forEach (key) ->
                 params = params.concat asRaw(mohair._with[key]).params()
 
-        params = params.concat that._params if that._params?
+        that._selects.forEach (s) ->
+            if isRaw s
+                params = params.concat s.params()
+            else if 'object' is typeof s
+                keys = Object.keys s
+                if keys.length is 0
+                    throw new Error 'select object must have at least one property'
+                keys.forEach (key) ->
+                    params = params.concat asRaw(s[key]).params()
 
         mohair._joins.forEach (join) ->
             if join.criterion?
@@ -113,15 +150,21 @@ select =
         params.push mohair._offset if mohair._offset?
         params
 
-module.exports.select = (sql, params) ->
+module.exports.select = ->
+    selects = Array.prototype.slice.call arguments
     object = Object.create select
-    object._sql = sql
-    object._params = params
+    if selects.length is 0
+        selects = ['*']
+    object._selects = selects
     object
 
 update =
     sql: (mohair) ->
         that = this
+
+        unless mohair._table?
+            throw new Error 'sql of update requires call to table before it'
+
         table = mohair._escape mohair._table
         keys = Object.keys that._updates
 
@@ -153,6 +196,10 @@ module.exports.update = (updates) ->
 deletePrototype =
     sql: (mohair) ->
         that = this
+
+        unless mohair._table?
+            throw new Error 'sql of delete requires call to table before it'
+
         table = mohair._escape mohair._table
         sql = "DELETE FROM #{table}"
         sql += " WHERE #{mohair._where.sql()}" if mohair._where?
